whir: unify utils in a single file
