+++ src/bin/benchmark.rs
@@ -27,8 +27,8 @@ use whir::{
     poly_utils::{coeffs::CoefficientList, multilinear::MultilinearPoint},
     whir::{
         domainsep::DigestDomainSeparator,
-        fs_utils::{DigestToUnitDeserialize, DigestToUnitSerialize},
         statement::{Statement, StatementVerifier, Weights},
+        utils::{DigestToUnitDeserialize, DigestToUnitSerialize},
     },
 };
