+++ src/bin/main.rs
@@ -22,8 +22,8 @@ use whir::{
     poly_utils::{coeffs::CoefficientList, evals::EvaluationsList, multilinear::MultilinearPoint},
     whir::{
         domainsep::DigestDomainSeparator,
-        fs_utils::{DigestToUnitDeserialize, DigestToUnitSerialize},
         statement::{Statement, StatementVerifier, Weights},
+        utils::{DigestToUnitDeserialize, DigestToUnitSerialize},
     },
 };
