+++ src/crypto/merkle_tree/blake3.rs
@@ -16,7 +16,7 @@ use spongefish::{
 use super::{HashCounter, IdentityDigestConverter};
 use crate::whir::{
     domainsep::DigestDomainSeparator,
-    fs_utils::{DigestToUnitDeserialize, DigestToUnitSerialize},
+    utils::{DigestToUnitDeserialize, DigestToUnitSerialize},
 };
 
 #[derive(
