+++ src/crypto/merkle_tree/keccak.rs
@@ -17,7 +17,7 @@ use spongefish::{
 use super::{HashCounter, IdentityDigestConverter};
 use crate::whir::{
     domainsep::DigestDomainSeparator,
-    fs_utils::{DigestToUnitDeserialize, DigestToUnitSerialize},
+    utils::{DigestToUnitDeserialize, DigestToUnitSerialize},
 };
 
 #[derive(
