+++ src/whir/verifier.rs
@@ -19,7 +19,7 @@ use crate::{
     poly_utils::{coeffs::CoefficientList, multilinear::MultilinearPoint},
     sumcheck::SumcheckPolynomial,
     utils::expand_randomness,
-    whir::fs_utils::{get_challenge_stir_queries, DigestToUnitDeserialize},
+    whir::utils::{get_challenge_stir_queries, DigestToUnitDeserialize},
 };
 
 pub struct Verifier<F, MerkleConfig, PowStrategy>
